All specs implemented. Test file complete; tests all specs.
Newline character at the end of the file test has one odd case. If only a single newline is there, the Scanner does not recognize it, and will print that a newline character is needed. It recognizes newlines at all other points of the file, and it sees one if there are two or more newlines at the end of the file, but not just one. Opening the test file in a hex editor while it contained a single newline at the end showed a "0d 0a" to denote that a newline was there, but the hasNextLine method from Scanner did not see that as being another line.